Refactor: Main 페이지 작업 Type 정리
